following system added

+++ app/controllers/static_pages_controller.rb
@@ -2,11 +2,14 @@ class StaticPagesController < ApplicationController
   before_action :url_log, only: [:home, :about, :help, :contact]
 
   def home
-    session[:tweet_count] = 20 if session[:tweet_count].nil?
-    @tweets = Tweet.all.order(:created_at).reverse.take(session[:tweet_count])
-    @tweet = Tweet.new()
-    @user = log_user
-    session.delete(:tweet_count)
+    if logged_in?
+      session[:tweet_count] = 20 if session[:tweet_count].nil?
+      @user = log_user
+      @tweets = @user.feed.order(:created_at).reverse.take(session[:tweet_count])
+      @tweet = Tweet.new()
+      session.delete(:tweet_count)
+    end
+
   end
 
   def reload
@@ -15,9 +18,9 @@ class StaticPagesController < ApplicationController
       format.html { redirect_back_to(root_url) }
       format.js {
         case session[:forwarding_url]
-        when root_url
+        when root_url(params[:id])
           @user = log_user
-          @tweets = Tweet.all.order(:created_at).reverse.take(session[:tweet_count])
+          @tweets = @user.feed.order(:created_at).reverse.take(session[:tweet_count])
         when user_url(params[:id])
           @user = User.find(params[:id])
           @tweets = @user.tweets.order(:created_at).reverse.take(session[:tweet_count])
@@ -35,5 +38,5 @@ class StaticPagesController < ApplicationController
 
   def contact
   end
-  
+
 end
